+++ backend/src/main/java/com/lecfantasy/backend/dto/JugadorEstadisticaDTO.java
@@ -14,5 +14,5 @@ public class JugadorEstadisticaDTO {
     private int deaths;
     private int assists;
     private int cs;
-    private Double puntosGenerados;
+    private Long puntosGenerados;
 }
